# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Restaurant API",
    "version": "1.0.0",
    "description": "API for managing restaurant operations"
  },
  "servers": [
    {
      "url": "http://localhost:8888",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/reservations": {
      "get": {
        "tags": ["Reservations"],
        "summary": "Reservations",
        "description": "The authenticated and authorized user can view reservations",
        "responses": {
          "200": {
            "description": "Reservations.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userId": {
                            "type": "object",
                            "properties": {
                              "username": {
                                "type": "string"
                              },
                              "email": {
                                "type": "string"
                              },
                              "phone": {
                                "type": "string"
                              }
                            }
                          },
                          "date": {
                            "type": "string"
                          },
                          "time": {
                            "type": "string"
                          },
                          "guests": {
                            "type": "number"
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "You dont have permission."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "post": {
        "tags": ["Reservations"],
        "summary": "Make reservations",
        "description": "The authenticated user can make a reservations.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string"
                  },
                  "time": {
                    "type": "string"
                  },
                  "guests": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reservations created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userId": {
                            "type": "object",
                            "properties": {
                              "username": {
                                "type": "string"
                              },
                              "email": {
                                "type": "string"
                              },
                              "phone": {
                                "type": "string"
                              }
                            }
                          },
                          "date": {
                            "type": "string"
                          },
                          "time": {
                            "type": "string"
                          },
                          "guests": {
                            "type": "number"
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/api/reservations/user": {
      "get": {
        "tags": ["Reservations"],
        "summary": "Reservations",
        "description": "The authenticated user can view his reservations",
        "responses": {
          "200": {
            "description": "Reservations.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userId": {
                            "type": "object",
                            "properties": {
                              "username": {
                                "type": "string"
                              },
                              "email": {
                                "type": "string"
                              },
                              "phone": {
                                "type": "string"
                              }
                            }
                          },
                          "date": {
                            "type": "string"
                          },
                          "time": {
                            "type": "string"
                          },
                          "guests": {
                            "type": "number"
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/api/reservations/{id}": {
      "put": {
        "tags": ["Reservations"],
        "summary": "Reservation update",
        "description": "The authenticated user can update his reservation.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Reservation id."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string"
                  },
                  "time": {
                    "type": "string"
                  },
                  "guests": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Reservation updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "delete": {
        "tags": ["Reservations"],
        "summary": "Delete Reservation",
        "description": "The authenticated user can cancel the reservation.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Reservation id."
          }
        ],
        "responses": {
          "200": {
            "description": "Reservation cancelled successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
  }
}